+++ template/framework/preact-ts/src/vite-env.d.ts
@@ -1,2 +1,2 @@
 /// <reference types="vite/client" />
-/// <reference types="@samrum/vite-plugin-web-extension/client" />
+/// <reference types="@ugurkellecioglu/vite-plugin-web-extension/client" />