+++ template/framework/svelte/vite.config.js
@@ -1,6 +1,6 @@
 import { defineConfig } from "vite";
 import { svelte } from "@sveltejs/vite-plugin-svelte";
-import webExtension from "@samrum/vite-plugin-web-extension";
+import webExtension from "@ugurkellecioglu/vite-plugin-web-extension";
 import { getManifest } from "./src/manifest.js";
 
 // https://vitejs.dev/config/