+++ template/framework/vanilla/vite.config.js
@@ -1,5 +1,5 @@
 import { defineConfig } from "vite";
-import webExtension from "@samrum/vite-plugin-web-extension";
+import webExtension from "@ugurkellecioglu/vite-plugin-web-extension";
 import path from "path";
 import { getManifest } from "./src/manifest";