+++ utils/banner.js
@@ -1,3 +1,3 @@
-const banner = "@samrum/create-vite-plugin-web-extension";
+const banner = "@ugurkellecioglu/create-vite-plugin-web-extension";
 
 export default banner;